Makefile: Fix DWC_LIB_DIR for rootfs images from buildroot
When we're building usb-scripts using buildroot, we need to specify the
libdir for the lib scripts. With the new change to move all the lib
scripts into <install>/lib directory, it did not account for "libdir"
setting.

This patch properly set the DWC_LIB_DIR to the correct install location.

Change-Id: I587feca92bb8033cab31cb9c4c441e7be9af7424
Signed-off-by: Thinh Nguyen <thinhn@synopsys.com>