Introduce lndir in Makefile for buildroot
Link to the lib scripts when building using buildroot will not link
correctly to the rootfs because it takes the absolute path that includes
the buildroot direction and the location where the rootfs is. Let's add
an option for the user to specify where the link for the lib scripts
should be for the lib scripts.

Change-Id: Idc34a21aad4aaf8dff37b7e43aac9e6e34b0e6d1
Signed-off-by: Thinh Nguyen <thinhn@synopsys.com>